Refactor get_pretty_string method in Term class for subclass customization
